Fix build fail on Windows due to C++11 usage
ggstat does not compile on Windows with R 3.3.0 and Rtools 3.3, because `isnan` in frange.cpp is C++11, and that does not seem to work (it returns the error `'isnan' was not declared in this scope`). Build on Ubuntu however runs without problems. I changed `isnan` to `ISNAN` as in Group-1d.h